refactor(ui): vanilla-extract api를 사용하지 않은 파일은 css.ts -> ts로 변경
